fix(Impl): fix invalid precondition check in RestTransport.Query, increase ScoredVector.Score precision, extend examples in playground

+++ src/VectorTypes.cs
@@ -21,7 +21,7 @@ public readonly record struct SparseVector
 public record ScoredVector
 {
     public required string Id { get; init; }
-    public required float Score { get; init; }
+    public required double Score { get; init; }
     public float[]? Values { get; init; }
     public SparseVector? SparseValues { get; init; }
     public MetadataMap? Metadata { get; init; }
